Removed comments + added gitignoe
I initially put in the comments to help my thought process and to serve
as a reminder to my future self why I did what I was doing.

Since those notes are archived on Github, they're unnecessary in future
versions.

I've also added a gitignore to the ListOfFollowers.txt. Since one is
created for you on use and I don't want to accidentally push my
ListOfFollowers.

In the future, a .ini will be added to this gitignore as well for the
user to input their Client-ID.