Variable name updates
SECONDS to SECONDSTOSLEEP - Making this more explicit. The pgoram sleeps
2 seconds to avoid Twitch API's rate limiting.

loginURL to userLoginEndpoint - We're calling on users with a login name
parameter. Buy inputting a plain text username, we are setting up the
login user login endpoint for calls.

loginURLResponse to userLoginResponse - We are storing the response of
the userLoginEndpoint call in this response.

userLoginID - The variable didn't update but this is now made more clear
what it stores. From the userLoginResponse, we are grabbing the ID of
the plaintext login name and storing it here.

loginURL to userFollowersEndpoint - We are calling on users to find out
who follows the userLoginID we have stored.

loginResponse to userFollowersResponse - We are storing the response to
the userFollowersEndpoint call.